Reinstall Colima dependencies when restoring cache (#8)
Caching & restoring QEMU is complicated because it has many other dependencies. 

With some improved bash-fu we could cache them all, but this is an improvement for the future. 

So we will spend some extra time to reinstall all dependencies from Colima that are missing. This will ensure everything needed is properly installed.