Improve resliency of action (#19)
Given all the recent trouble with Homebrew's conflicts and the QEMU bottle, I'm trying to improve on this regard by doing the following:

Download Colima and Lima binaries directly from their Github releases. For now, always download the latest version. Later I might add inputs to pick the version. They don't need to be cached anymore.

By default, use brew only to install Docker client and Docker Compose.

A check for QEMU's brew bottle 8.1.0 has been added. This version is broken and needs to be reinstalled by the action for things to work.

Added an input that can be used to force a QEMU upgrade.