Fix Colima VM SSH hang issue by pinning QEMU to 9.0.2 (#40)
It seems like the most recent version of QEMU is causing the SSH server inside the Colima VM to not start, which hangs all builds (see #37). I'm not exactly sure why and debugging the situation inside the action runners is kinda difficult.

With this PR I am pinning QEMU to 9.0.2, which seems to work on the runner image `macos-13`. 

Unfortunately I couldn't make it work on the runner image `macos-12`. As this is a very old macOS version, I'm dropping support for it. If one day things get fixed, I might readd it as supported.