Add self test to cover access via VM network address  (#57)
The big important details here is the LNP (local network policy) that Apple is including by default on macOS 15 and beyond: by default, only root processes can access local network. Any non-root process that attempts to do so triggers a permission pop-up which cannot be accepted on environments without a GUI (i.e. Github Actions). 

This means that if you want a process inside a macOS 15 Github Action Runner to access the Colima VM through the VM's IP, you will need to run it as `root` or with `sudo`. 

This is documented in the README and tested in the self-test action in this repository.

This work and investigation started from #56.